Adding failing int- and decimal-to-double tests
