--- test/FastExpressionCompiler.IssueTests/Issue159_NumericComparisonsWithConversions.cs
@@ -1,153 +1,0 @@
-﻿using System;
-using NUnit.Framework;
-
-#if LIGHT_EXPRESSION
-using static FastExpressionCompiler.LightExpression.Expression;
-namespace FastExpressionCompiler.LightExpression.UnitTests
-#else
-using System.Linq.Expressions;
-using static System.Linq.Expressions.Expression;
-namespace FastExpressionCompiler.UnitTests
-#endif
-{
-    public class Issue159_NumericComparisonsWithConversions
-    {
-        [Test]
-        public void UnsignedLongComparisonsWithConversionsShouldWork()
-        {
-            var ulongParameter = Parameter(typeof(ValueHolder<ulong>), "ulongValue");
-            var ulongValueProperty = Property(ulongParameter, "Value");
-
-            var intVariable = Variable(typeof(ValueHolder<int>), "intValue");
-            var intValueProperty = Property(intVariable, "Value");
-
-            var newIntHolder = Assign(intVariable, New(intVariable.Type));
-
-            var ulongGtOrEqualToIntMinValue = GreaterThanOrEqual(
-                Convert(ulongValueProperty, intValueProperty.Type),
-                Constant(int.MinValue));
-
-            var ulongLtOrEqualToIntMaxValue = LessThanOrEqual(
-                ulongValueProperty,
-                Convert(Constant(int.MaxValue), ulongValueProperty.Type));
-
-            var ulongIsInIntRange = AndAlso(ulongGtOrEqualToIntMinValue, ulongLtOrEqualToIntMaxValue);
-
-            var ulongAsInt = Convert(ulongValueProperty, typeof(int));
-            var defaultInt = Default(intValueProperty.Type);
-            var ulongValueOrDefault = Condition(ulongIsInIntRange, ulongAsInt, defaultInt);
-
-            var intValueAssignment = Assign(intValueProperty, ulongValueOrDefault);
-
-            var block = Block(
-                new[] { intVariable },
-                newIntHolder,
-                intValueAssignment,
-                intVariable);
-
-            var ulongValueOrDefaultLambda = Lambda<Func<ValueHolder<ulong>, ValueHolder<int>>>(
-                block,
-                ulongParameter);
-
-            var ulongValueOrDefaultFunc = ulongValueOrDefaultLambda.CompileFast();
-            var result = ulongValueOrDefaultFunc.Invoke(new ValueHolder<ulong> { Value = ulong.MaxValue });
-
-            Assert.AreEqual(default(int), result.Value);
-        }
-
-        [Test]
-        public void FloatComparisonsWithConversionsShouldWork()
-        {
-            var floatParameter = Parameter(typeof(ValueHolder<float>), "floatValue");
-            var floatValueProperty = Property(floatParameter, "Value");
-
-            var nullableShortVariable = Variable(typeof(ValueHolder<short?>), "nullableShort");
-            var nullableShortValueProperty = Property(nullableShortVariable, "Value");
-
-            var newShortHolder = Assign(nullableShortVariable, New(nullableShortVariable.Type));
-
-            var floatGtOrEqualToShortMinValue = GreaterThanOrEqual(
-                floatValueProperty,
-                Convert(Constant(short.MinValue), floatValueProperty.Type));
-
-            var floatLtOrEqualToShortMaxValue = LessThanOrEqual(
-                floatValueProperty,
-                Convert(Constant(short.MaxValue), floatValueProperty.Type));
-
-            var floatIsInShortRange = AndAlso(floatGtOrEqualToShortMinValue, floatLtOrEqualToShortMaxValue);
-
-            var floatAsNullableShort = Convert(floatValueProperty, nullableShortValueProperty.Type);
-            var defaultNullableShort = Default(nullableShortValueProperty.Type);
-            var floatValueOrDefault = Condition(floatIsInShortRange, floatAsNullableShort, defaultNullableShort);
-
-            var shortValueAssignment = Assign(nullableShortValueProperty, floatValueOrDefault);
-
-            var block = Block(
-                new[] { nullableShortVariable },
-                newShortHolder,
-                shortValueAssignment,
-                nullableShortVariable);
-
-            var floatValueOrDefaultLambda = Lambda<Func<ValueHolder<float>, ValueHolder<short?>>>(
-                block,
-                floatParameter);
-
-            var source = new ValueHolder<float> { Value = 532.00f };
-
-            var floatValueOrDefaultFunc = floatValueOrDefaultLambda.CompileFast();
-            var result = floatValueOrDefaultFunc.Invoke(source);
-
-            Assert.AreEqual((short)532, result.Value);
-        }
-
-        [Test]
-        public void FloatComparisonsWithConversionsShouldWork2()
-        {
-            var floatParameter = Parameter(typeof(ValueHolder<float>), "floatValue");
-            var floatValueProperty = Property(floatParameter, "Value");
-
-            var nullableShortVariable = Variable(typeof(ValueHolder<short>), "short");
-            var nullableShortValueProperty = Property(nullableShortVariable, "Value");
-
-            var newShortHolder = Assign(nullableShortVariable, New(nullableShortVariable.Type));
-
-            var floatGtOrEqualToShortMinValue = GreaterThanOrEqual(
-                floatValueProperty,
-                Convert(Constant(short.MinValue), floatValueProperty.Type));
-
-            var floatLtOrEqualToShortMaxValue = LessThanOrEqual(
-                floatValueProperty,
-                Convert(Constant(short.MaxValue), floatValueProperty.Type));
-
-            var floatIsInShortRange = AndAlso(floatGtOrEqualToShortMinValue, floatLtOrEqualToShortMaxValue);
-
-            var floatAsNullableShort = Convert(floatValueProperty, nullableShortValueProperty.Type);
-            var defaultNullableShort = Default(nullableShortValueProperty.Type);
-            var floatValueOrDefault = Condition(floatIsInShortRange, floatAsNullableShort, defaultNullableShort);
-
-            var shortValueAssignment = Assign(nullableShortValueProperty, floatValueOrDefault);
-
-            var block = Block(
-                new[] { nullableShortVariable },
-                newShortHolder,
-                shortValueAssignment,
-                nullableShortVariable);
-
-            var floatValueOrDefaultLambda = Lambda<Func<ValueHolder<float>, ValueHolder<short>>>(
-                block,
-                floatParameter);
-
-            var source = new ValueHolder<float> { Value = 532.00f };
-
-            var floatValueOrDefaultFunc = floatValueOrDefaultLambda.CompileFast();
-            var result = floatValueOrDefaultFunc.Invoke(source);
-
-            Assert.AreEqual((short)532, result.Value);
-        }
-
-        private class ValueHolder<T>
-        {
-            public T Value { get; set; }
-        }
-    }
-}
